undo components split
