--- src/components/Customer/MtoShipments/NTSDetailsForm.jsx
@@ -1,218 +1,0 @@
-import React, { Component } from 'react';
-import { func } from 'prop-types';
-import { get } from 'lodash';
-import { connect } from 'react-redux';
-import { Formik } from 'formik';
-import * as Yup from 'yup';
-import { Fieldset } from '@trussworks/react-uswds';
-
-import styles from './MtoShipmentFormStyles.module.scss';
-import { RequiredPlaceSchema } from './validationSchemas';
-import { PickupFields } from './FormGroups/PickupFields';
-
-import { TextInput } from 'components/form/fields';
-import { Form } from 'components/form/Form';
-import {
-  selectMTOShipmentForMTO,
-  createMTOShipment as createMTOShipmentAction,
-  updateMTOShipment as updateMTOShipmentAction,
-} from 'shared/Entities/modules/mtoShipments';
-import { selectServiceMemberFromLoggedInUser } from 'shared/Entities/modules/serviceMembers';
-import { showLoggedInUser as showLoggedInUserAction } from 'shared/Entities/modules/user';
-import { WizardPage } from 'shared/WizardPage';
-import { SHIPMENT_OPTIONS } from 'shared/constants';
-import { AddressShape } from 'types/address';
-import { NtsShipmentShape, WizardPageShape } from 'types/customerShapes';
-import { formatMtoShipment } from 'utils/formatMtoShipment';
-
-const NTSDetailsFormSchema = Yup.object().shape({
-  pickup: RequiredPlaceSchema,
-  customerRemarks: Yup.string(),
-});
-
-class NTSDetailsForm extends Component {
-  constructor(props) {
-    super(props);
-    const hasDeliveryAddress = get(props.mtoShipment, 'destinationAddress', false);
-    this.state = {
-      hasDeliveryAddress,
-      useCurrentResidence: false,
-      initialValues: {},
-    };
-  }
-
-  componentDidMount() {
-    const { showLoggedInUser } = this.props;
-    showLoggedInUser();
-  }
-
-  submitMTOShipment = ({ pickup, customerRemarks }) => {
-    const { createMTOShipment, wizardPage } = this.props;
-    const { moveId } = wizardPage.match.params;
-
-    const pendingMtoShipment = formatMtoShipment({
-      moveId,
-      pickup,
-      customerRemarks,
-      shipmentType: SHIPMENT_OPTIONS.NTS,
-    });
-
-    createMTOShipment(pendingMtoShipment);
-  };
-
-  handleChangeHasDeliveryAddress = () => {
-    this.setState((prevState) => {
-      return { hasDeliveryAddress: !prevState.hasDeliveryAddress };
-    });
-  };
-
-  // Use current residence
-  handleUseCurrentResidenceChange = (currentValues) => {
-    const { initialValues } = this.state;
-    const { currentResidence, wizardPage, mtoShipment } = this.props;
-    this.setState(
-      (state) => ({ useCurrentResidence: !state.useCurrentResidence }),
-      () => {
-        // eslint-disable-next-line react/destructuring-assignment
-        if (this.state.useCurrentResidence) {
-          this.setState({
-            initialValues: {
-              ...initialValues,
-              ...currentValues,
-              pickup: {
-                address: {
-                  street_address_1: currentResidence.street_address_1,
-                  street_address_2: currentResidence.street_address_2,
-                  city: currentResidence.city,
-                  state: currentResidence.state,
-                  postal_code: currentResidence.postal_code,
-                },
-              },
-            },
-          });
-        } else {
-          // eslint-disable-next-line no-lonely-if
-          if (wizardPage.match.params.moveId === initialValues.moveTaskOrderID) {
-            this.setState({
-              initialValues: {
-                ...initialValues,
-                ...currentValues,
-                pickup: {
-                  address: {
-                    street_address_1: mtoShipment.pickupAddress.street_address_1,
-                    street_address_2: mtoShipment.pickupAddress.street_address_2,
-                    city: mtoShipment.pickupAddress.city,
-                    state: mtoShipment.pickupAddress.state,
-                    postal_code: mtoShipment.pickupAddress.postal_code,
-                  },
-                },
-              },
-            });
-          } else {
-            this.setState({
-              initialValues: {
-                ...initialValues,
-                ...currentValues,
-                pickup: {
-                  address: {
-                    street_address_1: '',
-                    street_address_2: '',
-                    city: '',
-                    state: '',
-                    postal_code: '',
-                  },
-                },
-              },
-            });
-          }
-        }
-      },
-    );
-  };
-
-  render() {
-    const { wizardPage } = this.props;
-    const { pageKey, pageList, match, history } = wizardPage;
-    const { useCurrentResidence, initialValues } = this.state;
-    const fieldsetClasses = 'margin-top-2';
-    return (
-      <Formik
-        initialValues={initialValues}
-        enableReinitialize
-        validateOnBlur
-        validateOnChange
-        validationSchema={NTSDetailsFormSchema}
-      >
-        {({ values, dirty, isValid }) => (
-          <WizardPage
-            canMoveNext={dirty && isValid}
-            match={match}
-            pageKey={pageKey}
-            pageList={pageList}
-            push={history.push}
-            handleSubmit={() => this.submitMTOShipment(values, dirty)}
-          >
-            <h1>Now lets arrange details for the professional movers</h1>
-            <Form className={styles.HHGDetailsForm}>
-              <PickupFields
-                fieldsetClasses={fieldsetClasses}
-                useCurrentResidence={useCurrentResidence}
-                onCurrentResidenceChange={this.handleUseCurrentResidenceChange}
-                values={values.pickup}
-              />
-              <Fieldset legend="Remarks" className={fieldsetClasses}>
-                <TextInput
-                  label="Anything else you would like us to know?"
-                  labelHint="(optional)"
-                  data-testid="remarks"
-                  name="customerRemarks"
-                  id="customerRemarks"
-                  maxLength={1500}
-                  value={values.customerRemarks}
-                />
-              </Fieldset>
-            </Form>
-          </WizardPage>
-        )}
-      </Formik>
-    );
-  }
-}
-
-NTSDetailsForm.propTypes = {
-  wizardPage: WizardPageShape,
-  createMTOShipment: func.isRequired,
-  showLoggedInUser: func.isRequired,
-  currentResidence: AddressShape.isRequired,
-  mtoShipment: NtsShipmentShape,
-};
-
-NTSDetailsForm.defaultProps = {
-  wizardPage: {
-    pageList: [],
-    pageKey: '',
-    match: { isExact: false, params: { moveID: '' } },
-  },
-  mtoShipment: {
-    id: '',
-    customerRemarks: '',
-    requestedPickupDate: '',
-  },
-};
-
-const mapStateToProps = (state, ownProps) => {
-  const props = {
-    mtoShipment: selectMTOShipmentForMTO(state, ownProps.wizardPage.match.params.moveId),
-    currentResidence: get(selectServiceMemberFromLoggedInUser(state), 'residential_address', {}),
-  };
-  return props;
-};
-
-const mapDispatchToProps = {
-  createMTOShipment: createMTOShipmentAction,
-  updateMTOShipment: updateMTOShipmentAction,
-  showLoggedInUser: showLoggedInUserAction,
-};
-
-export { NTSDetailsForm as NTSDetailsFormComponent };
-export default connect(mapStateToProps, mapDispatchToProps)(NTSDetailsForm);
